Projects render condition changed
