+++ src/components/Button/Button.types.tsx
@@ -2,7 +2,6 @@ import { HTMLAttributes, PropsWithChildren } from 'react';
 
 import { PaletteColorType } from '@/types/shared.types';
 
-
 const buttonVariants = ['primary', 'secondary', 'toggle', 'filter', 'filterStatusOption'] as const;
 
 export interface ButtonProps extends HTMLAttributes<HTMLButtonElement>, PropsWithChildren {
