+++ src/pages/index.tsx
@@ -1,7 +1,6 @@
 import Head from 'next/head';
 
 import { FormatE1, Header, Partnership, Projects, WhoWeAre } from '@/containers/HomePage';
-import { ProjectContextProvider } from '@/context/projectPage.context';
 
 const Home = () => (
   <>
@@ -13,9 +12,7 @@ const Home = () => (
     </Head>
     <Header />
     <WhoWeAre />
-    <ProjectContextProvider>
       <Projects />
-    </ProjectContextProvider>
     <FormatE1 />
     <Partnership />
   </>
